scripts/requirements.txt
anthropic
beautifulsoup4
fuzzywuzzy
openai
python-dotenv
python-Levenshtein
supabase
